make the AllFavoritesList have a generic onEvent method
