+++ app/src/main/kotlin/org/cru/godtools/ui/dashboard/home/HomeLayout.kt
@@ -5,8 +5,6 @@ import androidx.compose.animation.fadeIn
 import androidx.compose.animation.fadeOut
 import androidx.compose.foundation.ExperimentalFoundationApi
 import androidx.compose.foundation.clickable
-import androidx.compose.foundation.interaction.DragInteraction
-import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
@@ -37,10 +35,6 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.viewmodel.compose.viewModel
 import java.util.Locale
-import org.burnoutcrew.reorderable.ReorderableItem
-import org.burnoutcrew.reorderable.detectReorderAfterLongPress
-import org.burnoutcrew.reorderable.rememberReorderableLazyListState
-import org.burnoutcrew.reorderable.reorderable
 import org.cru.godtools.BuildConfig
 import org.cru.godtools.R
 import org.cru.godtools.analytics.model.OpenAnalyticsActionEvent.Companion.ACTION_OPEN_LESSON
@@ -53,7 +47,6 @@ import org.cru.godtools.ui.banner.Banners
 import org.cru.godtools.ui.tools.LessonToolCard
 import org.cru.godtools.ui.tools.PreloadTool
 import org.cru.godtools.ui.tools.SquareToolCard
-import org.cru.godtools.ui.tools.ToolCard
 import org.cru.godtools.ui.tools.ToolCardEvent
 
 private val PADDING_HORIZONTAL = 16.dp
@@ -291,90 +284,6 @@ private fun NoFavoriteTools(
                 .align(Alignment.CenterHorizontally)
         ) {
             Text(stringResource(R.string.dashboard_home_section_favorites_action_all_tools))
-        }
-    }
-}
-
-@Composable
-@OptIn(ExperimentalFoundationApi::class)
-internal fun AllFavoritesList(
-    viewModel: HomeViewModel = viewModel(),
-    onOpenTool: (Tool?, Locale?, Locale?) -> Unit,
-    onOpenToolDetails: (String) -> Unit,
-) {
-    val favoriteTools by viewModel.reorderableFavoriteTools.collectAsState()
-
-    val reorderableState = rememberReorderableLazyListState(
-        // only support reordering tool items
-        canDragOver = { (_, k), _ -> (k as? String)?.startsWith("tool:") == true },
-        onMove = { from, to ->
-            val fromPos = favoriteTools.indexOfFirst { it.code == (from.key as? String)?.removePrefix("tool:") }
-            val toPos = favoriteTools.indexOfFirst { it.code == (to.key as? String)?.removePrefix("tool:") }
-            if (fromPos != -1 && toPos != -1) viewModel.moveFavoriteTool(fromPos, toPos)
-        },
-        onDragEnd = { _, _ -> viewModel.commitFavoriteToolOrder() }
-    )
-
-    LazyColumn(
-        contentPadding = PaddingValues(16.dp),
-        state = reorderableState.listState,
-        modifier = Modifier
-            .reorderable(reorderableState)
-            .detectReorderAfterLongPress(reorderableState)
-    ) {
-        item("header", "header") {
-            Text(
-                stringResource(R.string.dashboard_home_section_favorites_title),
-                style = MaterialTheme.typography.titleLarge,
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .animateItemPlacement()
-            )
-        }
-
-        items(favoriteTools, key = { "tool:${it.code}" }) { tool ->
-            ReorderableItem(reorderableState, "tool:${tool.code}") { isDragging ->
-                val interactionSource = remember { MutableInteractionSource() }
-                interactionSource.reorderableDragInteractions(isDragging)
-
-                PreloadTool(tool)
-                ToolCard(
-                    toolCode = tool.code.orEmpty(),
-                    confirmRemovalFromFavorites = true,
-                    interactionSource = interactionSource,
-                    onEvent = {
-                        when (it) {
-                            is ToolCardEvent.Click, is ToolCardEvent.OpenTool -> {
-                                viewModel.recordOpenClickInAnalytics(ACTION_OPEN_TOOL, it.tool?.code, SOURCE_FAVORITE)
-                                onOpenTool(it.tool, it.lang1, it.lang2)
-                            }
-                            is ToolCardEvent.OpenToolDetails -> {
-                                it.tool?.code?.let {
-                                    viewModel.recordOpenClickInAnalytics(ACTION_OPEN_TOOL_DETAILS, it, SOURCE_FAVORITE)
-                                    onOpenToolDetails(it)
-                                }
-                            }
-                        }
-                    },
-                    modifier = Modifier
-                        .padding(top = 16.dp)
-                        .fillMaxWidth()
-                )
-            }
-        }
-    }
-}
-
-@Composable
-private fun MutableInteractionSource.reorderableDragInteractions(isDragging: Boolean) {
-    val dragState = remember { object { var start: DragInteraction.Start? = null } }
-    LaunchedEffect(isDragging) {
-        when (val start = dragState.start) {
-            null -> if (isDragging) dragState.start = DragInteraction.Start().also { emit(it) }
-            else -> if (!isDragging) {
-                dragState.start = null
-                emit(DragInteraction.Stop(start))
-            }
         }
     }
 }
